Remove `IsFirstAttempt` flags on requests.
These flags existed as an optimization that allowed us to bypass idempotency
checks if we knew that the request had never been made before. However, this
approach assumes that the requests must arrive in order, which cannot be
guaranteed as we expand Veracity to work in a cluster of multiple nodes that
make requests to each other via an internal gRPC API, because each request is
handled in a separate goroutine.