Remove `Exchange.Context`.
Including the context in exchange in the context gives the sending side control of the
receiving side's behavior in a confusing way. Is it the sender or the
receiver's responsibility to handle timeouts, for example?

By removing the context from the exchange it is only available to the sender, eliminating any confusion.
The receiver _must_ provide a response, even if the sender has stopped
listening for it.